Backend/requirements.txt
fastapi==0.135.3
uvicorn==0.43.0
sqlalchemy==2.0.49
asyncpg==0.31.0
psycopg2-binary==2.9.11
alembic==1.18.4
python-jose==3.5.0
passlib==1.7.4
bcrypt==5.0.0
pydantic==2.12.5
pydantic-settings==2.13.1
python-dotenv==1.2.2
python-multipart==0.0.24
cryptography==46.0.6
email-validator>=2.0.0
